requirements.txt
# requirements.txt
torch
torchvision
timm
streamlit
opencv-python
albumentations
scikit-learn
pandas
numpy
matplotlib
plotly
Pillow
tqdm
fpdf
seaborn
grad-cam